RLM Docker
Dockerfile builds a basic Rocky 9 image with the RLM server included.

docker-compose.yaml contains build commands as well as run commands. This image will only build on amd64 as that’s the only architecture the server is available for.

WIP - container needs to expose ports, deal with MAC address, etc.