RLM v16.1
Moved to RLM v16.1, downloaded directly from the Reprise website, and added genarts.set to the repo so there is no dependency on the Boris FX RLM installer.

RLM 16 defaults to using https so there is now support for providing an SSL certificate, at `certs/rlm-cert.pem` and `certs/rlm-key.prm`. If these files are not provided, a self-signed certficate will be created on first run.

RLM now runs as an unprivileged user. I selected uid 100001 to get out of the normal user range.

The docker-compose file now specifies the runtime platform `linux/amd64` as well as this being specified in the bake section. This ensures that the image runs under emulation on arm64 hosts, which is required as genarts.set does not include support for the arm64_l1 platform.